Added support for high-voltage reconnect threshold
Added a new register REG_VSENSA_ON_THRESHOLD, which defaults to '0'. Modified the doSecondWork() function to check if the VSENSA voltage is greater than or equal to the ON threshold value in the code block that handles the power off state and an expired 'OFF_REMAINIG' timer.

Succesfully tested that I can set/det this register but then bricked my pidog when trying to reprogram it while also running the example.py script. Waiting on an AVR prgrammer to hopefully recover it.