Pidog unexpectedly cuts power to rpi
In discussions with David, we think the ADC reads for vsensa are occasionally returning garbage that is interpreted as voltage levels below the cutoff threshold. To remedy this, I've implemented his ema filter to hopefully prevent the supposed garbage data from causing unnecessary power cuts. I've also added some debug statements in the target code areas to print the values of the registers that are leading to this code path.